Horizon expects Quantum plugin to have floating IP support, however restproxy plugin does not currently have that support. Hence Horizon breaks.
Putting a temporary fix such that the Horizon will be able to make the
floating IP related calls to Quantum.

Note that this does introduce floating IP, this is just to get Horizon
working. You will notice the UI to associate floating IPs from Horizon,
however that will not be functional.