Firewall as a Service (FWaaS) APIs and DB Model
Implements: blueprint quantum-fwaas

blueprint: quantum-fwaas-plugin

This is the first iteration of the FWaaS implementation and
is geared towards implementing the model that will be
required to at least address the reference implementation.

This iteration will not include implementation of the following
features:
* grouping or dynamic objects
* application/service objects

Unit test coverage is not complete and tests are being constantly
added.

Change-Id: I57a62d6e9d3f1e6c4dd44cd5c745710a3d9e488e