ServiceInstanceInfo DB table
The table is part of the service insertion implementation.
The table is populated by each service plugin when a new instance
is created. When a new service_interface is created on a service
instance, the table is looked up to find the type of the service
for the instance.

Implements: blueprint service-base-class-and-insertion
Change-Id: Ie8015e16434445830b4df12a777f19e9fced391f