Don't update port with host id of None
In the L3 RPC code if the host for a port is not
present, it ends up calling update_port with the
host_id set to None. This does not update the host
id at all which leads to the same thing happening
on the next iteration. These pointless update calls
are expensive because the involve a semaphore and
calls to mechanism drivers.

This patch adjusts the logic to only send a port
update if it actually has a host to ensure is on
the port.

Change-Id: Ic55496dd2ba3abcef0a2de9fc8699c391b79fa51
Partial-Bug: #1445412
(cherry picked from commit 0155d1a172a30257a022a11209d0927a78c6d80a)